requirements.txt
discord.py[voice]==2.5.2
python-dotenv==1.1.1
yt_dlp==2025.06.30
